requirement.txt
asgiref==3.4.1
Django==3.2.10
django-mdeditor==0.1.18
importlib-metadata==4.9.0
Markdown==3.3.6
mysqlclient==2.1.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
zipp==3.6.0
